update project with new changes

+++ server/Controllers/CleanerController.js
@@ -1,4 +1,27 @@
 const CleanerModule = require("../Modules/CleanerModule");
+//בפונקציית יצירת מנקה חדש : יוצרים מנקה ומיד מעדכנים את היוזר שלו להכיל את הרפרנס אליו.
+const { updateUserReference } = require("./UserController");
+const { deleteUserFromQueueByIdFromLocal } = require("./ManagmerController");
+const { UserStatus } = require("../Constants/enums");
+const { default: mongoose } = require("mongoose");
+async function create(req, res) {
+    //Status מגיעה יחד עם שאר הפרטים של הזקן למרות שלא חלק מהאישי-הוא נלקח הצד לקוח מהסלייס יוזר
+    if (req.body.Status != UserStatus.CONFIRMED) {
+        return res.status(500).send({ message: "The status of the cleaner must be CONFIRMED" });
+    }
+    else {
+        req.body.dayInWork = JSON.parse(req.body.dayInWork);
+        console.log("req.body  : ", req.body);
+        let c = new CleanerModule(req.body);
+
+        await c.save();
+        //עידכון הרפרנס של הזקן ביוזר שלו.
+        await updateUserReference(c._id, c.RefId);
+        // מוציאים אותו מהתור של המנהל ומעדכנים את הסטטוס להיות פעיל
+        deleteUserFromQueueByIdFromLocal(null, c._id);
+        res.status(200).send(c);
+    }
+}
 
 async function getAll(req, res) {
     let arrC = await CleanerModule.find();
@@ -8,28 +31,6 @@ async function getAll(req, res) {
 async function getById(req, res) {
     let c = await CleanerModule.findById(req.params.id);
     res.status(200).send(c);
-}
-//בפונקציית יצירת מנקה חדש : יוצרים מנקה ומיד מעדכנים את היוזר שלו להכיל את הרפרנס אליו.
-const {updateUserReference} = require("./UserController");
-const {deleteUserFromQueueByIdFromLocal} = require("./ManagmerController");
-const { UserStatus } = require("../Constants/enums");
-async function create(req, res) {
-   //Status מגיעה יחד עם שאר הפרטים של הזקן למרות שלא חלק מהאישי-הוא נלקח הצד לקוח מהסלייס יוזר
-    if (req.body.Status != UserStatus.CONFIRMED) {
-        return res.status(500).send({ message: "The status of the cleaner must be CONFIRMED" });
-    }
-    else{
-        req.body.dayInWork = JSON.parse(req.body.dayInWork);
-        console.log("req.body  : ", req.body);
-    let c = new CleanerModule(req.body);
-    
-    await c.save();
-     //עידכון הרפרנס של הזקן ביוזר שלו.
-     await updateUserReference(c._id,c.RefId);
-     // מוציאים אותו מהתור של המנהל ומעדכנים את הסטטוס להיות פעיל
-     deleteUserFromQueueByIdFromLocal(null,c._id);
-    res.status(200).send(c);
-    }
 }
 
 async function deleteById(req, res) {
@@ -42,4 +43,16 @@ async function update(req, res) {
     res.status(200).send(c);
 }
 
-module.exports = { getAll, getById, create, deleteById, update };
+async function getCleanerByUserId(req, res) {
+    try {
+        const { _id } = req.params;
+        const cleaner = await CleanerModule.findOne({ RefId: new mongoose.Types.ObjectId(_id) });
+        if (!cleaner) 
+            return res.status(404).send({ message: "cleaner not found..." });
+        res.status(200).send(cleaner);
+    } catch (err) {
+        res.status(500).send({ message: "Server error", error: err.message });
+    }
+}
+
+module.exports = { getAll, getById, create, deleteById, update, getCleanerByUserId };
