+++ server/Routers/CleanerRouter.js
@@ -1,13 +1,13 @@
 const express = require("express")
 const router = express.Router()
-const {getAll,getById,create,deleteById,update} = require("../Controllers/CleanerController")
-
-router.post("/create",create)
-router.get("/getAll",getAll)
-router.get("/getById/:id",getById)
-router.delete("/deleteById/:id",deleteById)
-router.put("/update/:id",update)
-
+const {getAll,getById,create,deleteById,update,getCleanerByUserId} = require("../Controllers/CleanerController")
+const{verifyToken}=require("../auth")
+router.post("/create",verifyToken,create)
+router.get("/getAll",verifyToken,getAll)
+router.get("/getById/:id",verifyToken,getById)
+router.delete("/deleteById/:id",verifyToken,deleteById)
+router.put("/update/:id",verifyToken,update)
+router.get("/getCleanerByUserId/:_id",verifyToken, getCleanerByUserId)
 module.exports = router
